+++ pynonymizer/database/mssql/__init__.py
@@ -18,6 +18,8 @@ _FAKE_COLUMN_TYPES = {
     FakeDataType.INT: "INT"
 }
 
+_LOCAL_SERVER = "(local)"
+_DEFAULT_PORT = "1433"
 
 class MsSqlProvider(DatabaseProvider):
     """
@@ -35,28 +37,21 @@ class MsSqlProvider(DatabaseProvider):
         # import here for fast-failiness
         import pyodbc
 
-        if db_host is not None:
-            raise DependencyError("db_host", "MsSqlProvider does not support remote servers due to backup file "
-                                             "location requirements. You must omit db_host from your configuration "
-                                             "and run pynonymizer on the same server as the database.")
-
         # TODO: The odbc port syntax doesn't seem to work with (local),1433 or port=1433
         # TODO: This needs attention, but as it's backwards compatible, we're going to disallow it here.
         if db_port is not None:
             raise DependencyError("db_port", "MsSqlProvider does not support custom ports. You must omit db_port "
                                              "from your configuration to continue.")
 
-        if driver is None:
-            self.__driver = self.__detect_driver()
-        else:
-            self.__driver = driver
-            
-        db_host = "(local)"
+        db_host = db_host or _LOCAL_SERVER
+        db_port = db_port or _DEFAULT_PORT
+        driver = driver or self.__detect_driver()
 
         super().__init__(db_host=db_host, db_user=db_user, db_pass=db_pass, db_name=db_name, db_port=db_port, seed_rows=seed_rows)
         self.__conn = None
         self.__db_conn = None
         self.__backup_compression = backup_compression
+        self.__driver = driver
 
     def __detect_driver(self):
         import pyodbc
@@ -69,6 +64,11 @@ class MsSqlProvider(DatabaseProvider):
 
         return ms_drivers[0]
 
+    def __require_local_server(self):
+        if self.db_host != _LOCAL_SERVER:
+            raise DependencyError("db_host", "This operation does not support remote servers due to backup file "
+                                                "location requirements. You must omit db_host from your configuration "
+                                                "and run pynonymizer on the same server as the database.")
 
     def __connection(self):
         import pyodbc
@@ -295,6 +295,8 @@ class MsSqlProvider(DatabaseProvider):
         self.__drop_seed_table()
 
     def restore_database(self, input_path):
+        self.__require_local_server()
+
         move_files = self.__get_file_moves(input_path)
 
         self.logger.info("Found %d files in %s", len(move_files), input_path )
@@ -310,6 +312,8 @@ class MsSqlProvider(DatabaseProvider):
         self.__async_operation_progress("Restoring Database", restore_cursor)
 
     def dump_database(self, output_path):
+        self.__require_local_server()
+
         with_options = []
         if self.__backup_compression:
             with_options.append("COMPRESSION")