+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Changed MSSQL connection behaviour: provider will now allow truncation using `ANSI_WARNINGS off;` when updating tables.
 - Fixed a typo in mssql, mysql and postgres when running strategy before/after scripts.
 - Changed verbose-mode behaviour to include more information via debug logs during CLI runs.
+- Added support for remote servers in MSSQL, provided RESTORE_DB/DUMP_DB is not being used. 
 
 ## [1.17.0] 2021-03-29
 - Added New _additional-opts_ style parameters which allow you to control the behaviour of the underlying dump and restore tools more effectively.